+++ example-search-service-core/src/main/java/org/goafabric/example/searchservice/crossfunctional/TenantIdBean.java
@@ -1,0 +1,14 @@
+package org.goafabric.example.searchservice.crossfunctional;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+
+@Component
+@Slf4j
+public class TenantIdBean {
+    public String getPrefix() {
+        final String prefix = "tenant-" + TenantIdInterceptor.getTenantId() + "-";
+        log.debug("got prefix " + prefix);
+        return prefix;
+    }
+}