+++ example-search-service-core/src/main/java/org/goafabric/example/searchservice/service/Person.java
@@ -1,0 +1,20 @@
+package org.goafabric.example.searchservice.service;
+
+import lombok.*;
+
+import java.util.List;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Person {
+    private String id;
+
+    private String firstName;
+    private String lastName;
+
+    private Address address;
+
+    private List<Skill> skills;
+}