+++ example-search-service-jpa/src/main/java/org/goafabric/example/searchservice/persistence/AddressBo.java
@@ -1,0 +1,24 @@
+package org.goafabric.example.searchservice.persistence;
+
+import lombok.Data;
+import org.hibernate.annotations.GenericGenerator;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+@Data
+@Entity
+@Table(name="address")
+public class AddressBo {
+    private static final long serialVersionUID = 1L;
+
+    @Id
+    @GeneratedValue(generator = "uuid")
+    @GenericGenerator(name = "uuid", strategy = "uuid2")
+    private String id;
+
+    private String street;
+    private String city;
+}